pyproject.toml
[project]
name = "DanaDiffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.3.2",
    "diffusers>=0.32.2",
    "dotenv>=0.9.9",
    "ipykernel>=6.29.5",
    "peft>=0.14.0",
    "perfusion-pytorch>=0.1.23",
    "transformers>=4.49.0",
]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
DanaDiffusion = { path = ".", editable = true }

[tool.pixi.dependencies]
sentencepiece = ">=0.2.0,<0.3"
protobuf = ">=5.29.3,<6"

